fixed all of my code

+++ Blatt 05/Blatt05/4-pfister-a3/4-pfister-a3.c
@@ -42,7 +42,7 @@ double subtract(){
     return a - b;
 }
 
-double multiply(){
+long double multiply(){
     printf("*** Sie befinden sich im Multiplizierer ***\n");
     double a, b = 0.0;
